fix: moved renderBlock to RenderUtils - removed the added line from MiscUtil
