+++ src/main/java/dev/ftb/mods/ftbstuffnthings/client/renders/AutoHammerRenderer.java
@@ -5,6 +5,7 @@ import com.mojang.blaze3d.vertex.SheetedDecalTextureGenerator;
 import com.mojang.blaze3d.vertex.VertexConsumer;
 import com.mojang.blaze3d.vertex.VertexMultiConsumer;
 import dev.ftb.mods.ftbstuffnthings.blocks.hammer.AutoHammerBlockEntity;
+import dev.ftb.mods.ftbstuffnthings.client.RenderUtil;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderer;
@@ -25,23 +26,8 @@ public class AutoHammerRenderer implements BlockEntityRenderer<AutoHammerBlockEn
         if (blockEntity.getLevel().isLoaded(blockEntity.getBlockPos())) {
             poseStack.pushPose();
             poseStack.translate(0.375, 0.3125, 0.375);
-            renderBlock(poseStack, bufferSource, packedLight, packedOverlay, blockEntity.getProcessingStack(), blockEntity.getDestroyStage());
+            RenderUtil.renderBlock(poseStack, bufferSource, packedLight, packedOverlay, blockEntity.getProcessingStack(), blockEntity.getDestroyStage());
             poseStack.popPose();
-        }
-    }
-
-    public static void renderBlock(PoseStack poseStack, MultiBufferSource bufferSource, int combinedLightIn, int combinedOverlayIn, ItemStack stack, int breakProgress) {
-        if (!stack.isEmpty() && stack.getItem() instanceof BlockItem blockItem) {
-            poseStack.scale(0.25f, 0.25f, 0.25f);
-
-            VertexConsumer vertexconsumer = new SheetedDecalTextureGenerator(bufferSource.getBuffer(ModelBakery.DESTROY_TYPES.get(breakProgress)), poseStack.last(), 1.0F);
-            MultiBufferSource bufferSource1 = type -> {
-                VertexConsumer vc = bufferSource.getBuffer(type);
-                return type.affectsCrumbling() ? VertexMultiConsumer.create(vertexconsumer, vc) : vc;
-            };
-
-            BlockState state = blockItem.getBlock().defaultBlockState();
-            Minecraft.getInstance().getBlockRenderer().renderSingleBlock(state, poseStack, bufferSource1, combinedLightIn, combinedOverlayIn, ModelData.EMPTY, null);
         }
     }
 }
